Move methods to ArtistsService
