+++ docker-compose.yml
@@ -8,10 +8,10 @@ services:
     restart: unless-stopped
     volumes:
       - ./Music:/music
-      - ./logs:/app/logs # Optional
+      - ./app:/app
       - ./tracking.yaml:/app/tracking.yaml:ro
-      - /etc/localtime:/etc/localtime:ro
     environment:
+      TZ: Europe/Madrid
       CRON_SCHEDULE: "0 0 * * *"
       CLIENT__ID: "" # Required
       CLIENT__SECRET: "" # Required
